requirements.txt
# Phish-Hook Requirements
# Machine Learning
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0

# Imbalanced Learning
imbalanced-learn>=0.8.0

# String Distance
python-Levenshtein>=0.12.0

# Certificate Transparency
certstream>=1.11

# Visualization (optional)
matplotlib>=3.4.0

# Utilities
python-dateutil>=2.8.0

